[refactor] attendance 패키지로 따로 분리
