+++ Backend/src/test/java/com/github/backend/service/attendance/impl/AttendanceServiceImplTest.java
@@ -1,4 +1,4 @@
-package com.github.backend.service.impl;
+package com.github.backend.service.attendance.impl;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
